third_party/blink/web_tests/accessibility/computed-role-expected.txt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS roleElement.computedRole is "alert"
PASS roleElement.computedRole is "alertdialog"
PASS roleElement.computedRole is "application"
PASS roleElement.computedRole is "article"
PASS implicitRoleElement.computedRole is "article"
PASS roleElement.computedRole is "banner"
PASS implicitRoleElement.computedRole is "banner"
PASS roleElement.computedRole is "button"
PASS implicitRoleElement.computedRole is "button"
PASS implicitRoleElement.computedRole is "button"
PASS roleElement.computedRole is "button"
PASS implicitRoleElement.computedRole is "button"
PASS implicitRoleElement.computedRole is "button"
PASS roleElement.computedRole is "complementary"
PASS implicitRoleElement.computedRole is "complementary"
PASS roleElement.computedRole is "contentinfo"
PASS implicitRoleElement.computedRole is "contentinfo"
PASS roleElement.computedRole is "checkbox"
PASS implicitRoleElement.computedRole is "checkbox"
PASS roleElement.computedRole is "dialog"
PASS implicitRoleElement.computedRole is "dialog"
PASS roleElement.computedRole is "grid"
PASS roleElement.computedRole is "row"
PASS roleElement.computedRole is "columnheader"
PASS roleElement.computedRole is "row"
PASS roleElement.computedRole is "gridcell"
PASS roleElement.computedRole is "combobox"
PASS implicitRoleElement.computedRole is "combobox"
PASS implicitRoleElement.computedRole is "combobox"
PASS roleElement.computedRole is "definition"
PASS roleElement.computedRole is "document"
PASS roleElement.computedRole is "form"
PASS implicitRoleElement.computedRole is "form"
PASS implicitRoleElement.computedRole is "form"
PASS implicitRoleElement.computedRole is "form"
PASS roleElement.computedRole is "generic"
PASS implicitRoleElement.computedRole is "generic"
PASS implicitRoleElement.computedRole is "generic"
PASS implicitRoleElement.computedRole is "generic"
PASS roleElement.computedRole is "group"
PASS implicitRoleElement.computedRole is "group"
PASS roleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS implicitRoleElement.computedRole is "heading"
PASS roleElement.computedRole is "link"
PASS implicitRoleElement.computedRole is "link"
PASS implicitRoleElement.computedRole is "link"
PASS roleElement.computedRole is "image"
PASS implicitRoleElement.computedRole is "image"
PASS roleElement.computedRole is "list"
PASS implicitRoleElement.computedRole is "list"
PASS implicitRoleElement.computedRole is "list"
PASS implicitRoleElement.computedRole is "list"
PASS implicitRoleElement.computedRole is "list"
PASS roleElement.computedRole is "listbox"
PASS roleElement.computedRole is "log"
PASS roleElement.computedRole is "main"
PASS implicitRoleElement.computedRole is "main"
PASS roleElement.computedRole is "marquee"
PASS roleElement.computedRole is "math"
PASS roleElement.computedRole is "menu"
PASS roleElement.computedRole is "menubar"
PASS roleElement.computedRole is "menuitem"
PASS roleElement.computedRole is "menuitemcheckbox"
PASS roleElement.computedRole is "menuitemradio"
PASS roleElement.computedRole is "note"
PASS roleElement.computedRole is "navigation"
PASS implicitRoleElement.computedRole is "navigation"
PASS roleElement.computedRole is "option"
PASS implicitRoleElement.computedRole is "option"
PASS implicitRoleElement.computedRole is "option"
PASS roleElement.computedRole is "none"
PASS roleElement.computedRole is "progressbar"
PASS implicitRoleElement.computedRole is "progressbar"
PASS roleElement.computedRole is "meter"
PASS implicitRoleElement.computedRole is "meter"
PASS roleElement.computedRole is "radio"
PASS implicitRoleElement.computedRole is "radio"
PASS roleElement.computedRole is "radiogroup"
PASS roleElement.computedRole is "region"
PASS implicitRoleElement.computedRole is "region"
PASS implicitRoleElement.computedRole is "region"
PASS implicitRoleElement.computedRole is "region"
PASS roleElement.computedRole is "scrollbar"
PASS roleElement.computedRole is "search"
PASS roleElement.computedRole is "separator"
PASS implicitRoleElement.computedRole is "separator"
PASS roleElement.computedRole is "slider"
PASS implicitRoleElement.computedRole is "slider"
PASS roleElement.computedRole is "spinbutton"
PASS implicitRoleElement.computedRole is "spinbutton"
PASS roleElement.computedRole is "status"
PASS implicitRoleElement.computedRole is "status"
PASS roleElement.computedRole is "tablist"
PASS roleElement.computedRole is "tab"
PASS roleElement.computedRole is "tabpanel"
PASS roleElement.computedRole is "textbox"
PASS implicitRoleElement.computedRole is "textbox"
PASS implicitRoleElement.computedRole is "textbox"
PASS implicitRoleElement.computedRole is "textbox"
PASS implicitRoleElement.computedRole is "textbox"
PASS implicitRoleElement.computedRole is "textbox"
PASS roleElement.computedRole is "timer"
PASS roleElement.computedRole is "toolbar"
PASS roleElement.computedRole is "tooltip"
PASS roleElement.computedRole is "tree"
PASS roleElement.computedRole is "treegrid"
PASS roleElement.computedRole is "treeitem"
PASS svgAElement.computedRole is "link"
PASS successfullyParsed is true

TEST COMPLETE
